feat(avatar): introduce filesystem caching
There is a noticeable delay waiting for these
images to be served. Adopting the filesystem in
an effort to speed things up, although python is still
the bottle neck for streaming the response back to the client.

We can iterate further if the perf is not where we want it to be.