feat(acl): introduce new category
As part of rolling out our Entitlements system to
control access to our Wiki instance we should ensure
the acl is extended to include a Software category for
our "Machines".

We may find that the term "Machines" becomes more abstract
over time, but revisiting this is out of scope until we have
more than the Wiki exception.