feat(api): use configured DEFAULT_LOCATION_ID for api transctions
I am not entirely happy with this implementation. I feel that this is
too far out. Rather than being handled on the API endpoint it should be
within the model method to prevent divergence of behaviour over time.

Unfortunately there is a circular dependency when importing the Location
model into the members/model file. This will take some more unpicking
and is likely better tackled as part of a larger refactoring effort.