feat: add location with checkin (#305)
* feat(members): support specifying Location with check-in

* feat(members): support checkin with location

* chore: drop commented out code

* chore: apply formatting

* sketch: support display location based check-ins

* feat(api): use configured DEFAULT_LOCATION_ID for api transctions

I am not entirely happy with this implementation. I feel that this is
too far out. Rather than being handled on the API endpoint it should be
within the model method to prevent divergence of behaviour over time.

Unfortunately there is a circular dependency when importing the Location
model into the members/model file. This will take some more unpicking
and is likely better tackled as part of a larger refactoring effort.

* chore: helper for checking out users

* fix(space-state): improved appearance for member listings

* test: update to align with new behaviour

* chore: add production configuration